Revert KeySelector
