+++ core/src/main/java/com/linecorp/armeria/client/circuitbreaker/CircuitBreakerMapping.java
@@ -19,6 +19,7 @@ package com.linecorp.armeria.client.circuitbreaker;
 import java.util.function.Function;
 
 import com.linecorp.armeria.client.ClientRequestContext;
+import com.linecorp.armeria.client.circuitbreaker.KeyedCircuitBreakerMapping.KeySelector;
 import com.linecorp.armeria.common.Request;
 
 /**
@@ -35,23 +36,12 @@ public interface CircuitBreakerMapping {
     }
 
     /**
-     * Returns a new {@link CircuitBreakerMapping} which maps {@link CircuitBreaker}s with the key selected by
-     * the specified {@link KeySelector}.
-     *
-     * @param keySelector the key selector
-     * @param factory the function that takes a method name and creates a new {@link CircuitBreaker}
-     */
-    static <K> CircuitBreakerMapping of(KeySelector<K> keySelector, Function<K, CircuitBreaker> factory) {
-        return new KeyedCircuitBreakerMapping<>(keySelector, factory);
-    }
-
-    /**
      * Creates a new {@link CircuitBreakerMapping} which maps {@link CircuitBreaker}s with method name.
      *
      * @param factory the function that takes a method name and creates a new {@link CircuitBreaker}
      */
     static CircuitBreakerMapping perMethod(Function<String, CircuitBreaker> factory) {
-        return new KeyedCircuitBreakerMapping<>(KeySelectorUtil.methodSelector, factory);
+        return new KeyedCircuitBreakerMapping<>(KeySelector.METHOD, factory);
     }
 
     /**
@@ -60,7 +50,7 @@ public interface CircuitBreakerMapping {
      * @param factory the function that takes a host name and creates a new {@link CircuitBreaker}
      */
     static CircuitBreakerMapping perHost(Function<String, CircuitBreaker> factory) {
-        return new KeyedCircuitBreakerMapping<>(KeySelectorUtil.hostSelector, factory);
+        return new KeyedCircuitBreakerMapping<>(KeySelector.HOST, factory);
     }
 
     /**
@@ -71,7 +61,7 @@ public interface CircuitBreakerMapping {
      *                creates a new {@link CircuitBreaker}
      */
     static CircuitBreakerMapping perHostAndMethod(Function<String, CircuitBreaker> factory) {
-        return new KeyedCircuitBreakerMapping<>(KeySelectorUtil.hostAndMethodSelector, factory);
+        return new KeyedCircuitBreakerMapping<>(KeySelector.HOST_AND_METHOD, factory);
     }
 
     /**
