--- core/src/main/java/com/linecorp/armeria/client/circuitbreaker/KeySelector.java
@@ -1,34 +1,0 @@
-/*
- * Copyright 2020 LINE Corporation
- *
- * LINE Corporation licenses this file to you under the Apache License,
- * version 2.0 (the "License"); you may not use this file except in compliance
- * with the License. You may obtain a copy of the License at:
- *
- *   https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
- * License for the specific language governing permissions and limitations
- * under the License.
- */
-package com.linecorp.armeria.client.circuitbreaker;
-
-import com.linecorp.armeria.client.ClientRequestContext;
-import com.linecorp.armeria.common.Request;
-import com.linecorp.armeria.common.RequestContext;
-
-/**
- * A key selector that returns the mapping key of the given {@link RequestContext} and {@link Request}.
- *
- * @param <K> the key type
- */
-@FunctionalInterface
-public interface KeySelector<K> {
-
-    /**
-     * Returns the mapping key of the given {@link RequestContext} and {@link Request}.
-     */
-    K get(ClientRequestContext ctx, Request req) throws Exception;
-}
